% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{arrange_dot_layout}
\alias{arrange_dot_layout}
\title{Arrange Dot Layout}
\usage{
arrange_dot_layout(portion, nrows, group, df = long_df)
}
\arguments{
\item{portion}{A numeric value specifying what proportion of a full circle should be used for drawing the plot.}

\item{nrows}{If \code{style = "dots"}, a numeric value indicating how many rows to use.}

\item{group}{The grouping of the data to be lapply-d to the function (i.e. to work out positions for each group on a separate graph)}
}
\description{
Function to calculate the position of polar co-ords for a dot layout parliament plot
}
