% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{calc_coordinates}
\alias{calc_coordinates}
\title{A function that calculates the coordinates of parliamentary seats in incomplete circular parliaments
E.g. The US (semicircle) and Australian (horsehoe) shaped parliaments}
\usage{
calc_coordinates(N, M, limits, segment = 0.5)
}
\arguments{
\item{N}{the total of number of seats}

\item{M}{the number of rows in parliament}

\item{limits}{the limits to seq the radii between- controls the 'shape' of the parliament}

\item{segment}{the percentage of a full circle for the final plot- defaults to 0.5 (a semicircle)}
}
\description{
A function that calculates the coordinates of parliamentary seats in incomplete circular parliaments
E.g. The US (semicircle) and Australian (horsehoe) shaped parliaments
}
\author{
Zoe Meers, Rob Hickman
}
