% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{combine_opposingbenches}
\alias{combine_opposingbenches}
\title{Combine left and right bench for opposing bench-style parliaments}
\usage{
combine_opposingbenches(left = NA, right = NA)
}
\arguments{
\item{left}{left hand side}

\item{right}{right hand side}
}
\description{
Combine left and right bench for opposing bench-style parliaments
}
\author{
Zoe Meers
}
