% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_legislaturelabels.R
\name{draw_legislaturelabels}
\alias{draw_legislaturelabels}
\title{Draw legislature labels}
\usage{
draw_legislaturelabels(total_parliamentary_seats = NULL,
  total_seats_label = TRUE, party_names = TRUE, seats_per_party = TRUE,
  type = c("horseshoe", "semicircle"), ...)
}
\arguments{
\item{party_names}{A logical variable for labelling different parties. Defaults to TRUE.}

\item{seats_per_party}{A logical variable for printing the number of seats per party next to the party name. Defaults to TRUE.}

\item{total_seats}{The number of total seats}
}
\description{
Draw legislature labels
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, party_names = data$party_short, parl_rows = 8, total_seats = sum(data$seats))
ggplot(usa_data, aes(x, y, color=party_long)) + geom_parliament_seats() + draw_legislaturelabels(n = 435, party_names = TRUE, seats_per_party = TRUE, type = 'semicircle')
}
\author{
Zoe Meers
}
