% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_majorityline.R
\name{draw_majoritythreshold}
\alias{draw_majoritythreshold}
\title{Draw majority line}
\usage{
draw_majoritythreshold(..., n = NULL, label = TRUE, type = c("horseshoe",
  "semicircle", "opposing_benches"))
}
\arguments{
\item{n}{The number of seats required for a majority}

\item{label}{A logical variable for labelling majority threshold. Defaults to TRUE.}

\item{type}{type of parliament (horseshoe, semicircle,opposing benches)}
}
\description{
Draw majority line
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, party_names = data$party_short, parl_rows = 8, total_seats = sum(data$seats))
ggplot(usa_data, aes(x, y, color=party_long)) + geom_parliament_seats() + draw_majorityline(n = 316, type = 'opposing_benches', label = FALSE)
}
\author{
Zoe Meers
}
