% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_legislaturelabels_DRAFT_2.R
\name{draw_majoritythreshold3}
\alias{draw_majoritythreshold3}
\title{Draw majority line}
\usage{
draw_majoritythreshold3(..., n = NULL, label = TRUE, linecolour = "black",
  linesize = 1, linetype = 1)
}
\arguments{
\item{n}{The number of seats required for a majority}

\item{label}{A logical variable for labelling majority threshold. Defaults to TRUE.}

\item{colour}{The colour of the majority line. Defaults to grey.}
}
\description{
Draw majority line
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, parl_rows = 8)
ggplot(usa_data, aes(x, y, color=party_long)) + geom_parliament_seats() + draw_majoritythreshold3(n = 218, label = FALSE, linecolour = "black") + theme_ggparliament()
}
\author{
Zoe Meers
}
