% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_legislaturelabels.R
\name{draw_totalseats}
\alias{draw_totalseats}
\title{Draw total number of seats in the middle of the parliament}
\usage{
draw_totalseats(n = NULL, inherit.aes = TRUE, size = 8,
  colour = "black", type = NA, ...)
}
\arguments{
\item{n}{The number of total seats}
}
\description{
Draw total number of seats in the middle of the parliament
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, parl_rows = 8)
ggplot(usa_data, aes(x, y, color=party_long)) + geom_parliament_seats() + draw_totalseats(n = 435, type = 'semicircle') + theme_void()
}
\author{
Zoe Meers
}
