% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_electoral_quota.R
\name{geom_electoral_quota}
\alias{geom_electoral_quota}
\title{Show electoral quotas
Define the quota variable in the function.}
\usage{
geom_electoral_quota(expr)
}
\arguments{
\item{expr}{Exp refer to the seats that fill an electoral quota.}
}
\description{
Show electoral quotas
Define the quota variable in the function.
}
\examples{
geom_electoral_quota(quota == 1)
}
\author{
Zoe Meers
}
