% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_hanging_seats}
\alias{geom_hanging_seats}
\title{Show hangins seats
Define hanging seats in the function}
\usage{
geom_hanging_seats(expr)
}
\description{
Show hangins seats
Define hanging seats in the function
}
\examples{
geom_hanging_seats(hanging_seats == 1)
}
\author{
Zoe Meers
}
