% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_highlight_government}
\alias{geom_highlight_government}
\title{Highlight governments
Define the government variable in the function.}
\source{
https://yutani.rbind.io/post/2017-11-07-ggplot-add/
}
\usage{
geom_highlight_government(expr)
}
\description{
Highlight governments
Define the government variable in the function.
}
\examples{
geom_highlight_parliament(government==1)
}
\author{
Zoe Meers
}
