% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_overhang_seats.R
\name{geom_overhang_seats}
\alias{geom_overhang_seats}
\title{Draw overhang seats in MMP electoral systems}
\usage{
geom_overhang_seats(expr)
}
\arguments{
\item{expr}{Exp refers to the designated overhang seats.}
}
\description{
Draw overhang seats in MMP electoral systems
}
\examples{
data <- ggparliament::election_data \%>\%
 filter(country == "Germany" 
 & year == "2013") \%>\% 
 mutate(seats = gsub("255", "311", seats)) \%>\%
 mutate(seats = as.numeric(as.character(seats))) \%>\%
 filter_all(all_vars(!grepl('Christian Social Union in Bavaria',.)))
overhangseats <- c(1, 0, 1, 0, 1, 0, 1, 0)
number_overhangseats <- c(16, 295, 11, 182, 3, 61, 3, 60)
german_data <- parliament_data(
 election_data = data,
 parl_rows = 11,
 party_seats = data$seats,
 type = "semicircle"
)
german_data$overhang_seats <- rep(overhangseats, number_overhangseats)
ggplot(german_data, aes(x,y,colour = party_short)) +
 geom_parliament_seats() +
 geom_overhang_seats(overhang_seats == 1) +
 theme_ggparliament
}
\author{
Zoe Meers
}
