% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_parliament_seats}
\alias{geom_parliament_seats}
\title{Parliament seats
The parliament seats geom is used for plotting data from parliament_data().}
\usage{
geom_parliament_seats(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{x}{the x coordinates}

\item{y}{the y coordinates}

\item{colour}{the colour variable}
}
\description{
Parliament seats
The parliament seats geom is used for plotting data from parliament_data().
}
\examples{
geom_parliament_seats(female == 1)
}
\author{
Zoe Meers
}
