% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_parliament_waffle}
\alias{geom_parliament_waffle}
\title{A ggplot2 geom for parliament plots}
\usage{
geom_parliament_waffle(total_seats = NULL, parl_rows = NULL,
  party_seats = NULL, size = NULL, party_names = NULL,
  type = "opposing_benches")
}
\arguments{
\item{total_seats}{the total number of seats in parliament}

\item{parl_rows}{number of rows in parliament}

\item{type}{type of parliament (horseshow, semicircle, circle, classroom, opposing benches)}

\item{data[[party_seats]]}{seats per party}

\item{data[[party_names]]}{names of political parties in parliament}
}
\description{
A ggplot2 geom for parliament plots
}
\author{
Zoe Meers
### ROB - not sure this needs to be here - does it add anything over using the square geom_point? ###
}
