% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_women_in_parliament}
\alias{geom_women_in_parliament}
\title{Highlight elected women
Define the "women" variable in the function.}
\usage{
geom_women_in_parliament(expr)
}
\description{
Highlight elected women
Define the "women" variable in the function.
}
\examples{
geom_women_in_parliament(female == 1)
}
\author{
Zoe Meers
}
