% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggparliament.R
\name{theme_ggparliament}
\alias{theme_ggparliament}
\title{A theme for ggparliament}
\usage{
theme_ggparliament()
}
\arguments{
\item{legend}{If legend = `TRUE`, add legend to plot. Defaults to `TRUE`. You can also write "top", "left", "right", "bottom" and the theme will plot the legend on the specified panel edge.}

\item{border}{If `TRUE` add panel border. Defaults to `FALSE`.}

\item{background_colour}{If background colour = `TRUE`, fill panel with a grey background. Defaults to `FALSE`.}
}
\description{
Calls the ggparliament theme. A reconstructed opinionated theme_void() ggplot2 theme.
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, parl_rows = 8)
ggplot(usa_data, aes(x, y, color = party_long)) + geom_parliament_seats() + geom_highlight_government(government == 1) + theme_ggparliament(legend = TRUE, background = TRUE, border = TRUE)
}
\author{
Zoe Meers
}
