% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_majoritythreshold.R
\name{draw_majoritythreshold}
\alias{draw_majoritythreshold}
\title{Draw majority line}
\usage{
draw_majoritythreshold(..., n = NULL, label = TRUE, type = NULL,
  colour = "grey")
}
\arguments{
\item{n}{The number of seats required for a majority}

\item{label}{A logical variable for labelling majority threshold. Defaults to TRUE.}

\item{type}{Type of parliament (horseshoe, semicircle,opposing benches)}

\item{colour}{The colour of the majority line. Defaults to grey.}
}
\description{
Draw majority line
}
\examples{
data <- ggparliament::election_data \%>\% filter(year == "2016" & country == "USA" & house == "Representatives")
usa_data <- parliament_data(election_data = data, type = "semicircle", party_seats = data$seats, parl_rows = 8)
ggplot(usa_data, aes(x, y, color=party_long)) + geom_parliament_seats() + draw_majoritythreshold(n = 316, label = FALSE, colour = "black", type = 'semicircle')
}
\author{
Zoe Meers
}
