% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_electoral_quota}
\alias{geom_electoral_quota}
\title{Show electoral quotas
Define the quota variable in the function.}
\usage{
geom_electoral_quota(expr)
}
\description{
Show electoral quotas
Define the quota variable in the function.
}
\examples{
geom_electoral_quota(quota == 1)
}
\author{
Zoe Meers
}
