% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_emphasize_parliamentarians}
\alias{geom_emphasize_parliamentarians}
\title{Highlight elected women
Define the "women" variable in the function.}
\usage{
geom_emphasize_parliamentarians(expr)
}
\description{
Highlight elected women
Define the "women" variable in the function.
}
\examples{
geom_emphasize_parliamentarians(female == 1)
}
\author{
Zoe Meers
}
