% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_overhang_seats}
\alias{geom_overhang_seats}
\title{Draw overhang seats in MMP electoral systems}
\usage{
geom_overhang_seats(expr)
}
\description{
Draw overhang seats in MMP electoral systems
}
\examples{
data <- ggparliament::election_data \%>\%
 filter(country == "Germany" & year == "2013") \%>\% # on the federal level, CSU is a part of CDU
 mutate(party_long = gsub("Christian Social Union in Bavaria", "Christian Democratic Union", party_long##' )) \%>\%
 mutate(party_short = gsub("CSU", "CDU", party_short)) \%>\%
 mutate(colour = gsub("#008AC5", "#000000", colour)) \%>\%
 mutate(seats = gsub("255", "311", seats)) \%>\%
 mutate(seats = as.numeric(as.character(seats))) \%>\%
 slice(1:4)
overhangseats <- c(1, 0, 1, 0, 1, 0, 1, 0)
number_overhangseats <- c(16, 295, 11, 182, 3, 61, 3, 60)
german_data <- parliament_data(
 election_data = data,
 parl_rows = 11,
 party_seats = data$seats,
 type = "semicircle"
)
german_data <- german_data \%>\% mutate(overhang_seats = rep(overhangseats, number_overhangseats))
ggplot(german_data, aes(x,y,colour = party_short)) +
 geom_parliament_seats() +
 geom_overhang_seats(overhang_seats == 1) +
 labs(colour = "Party", title = "German Bundestag - 2013 election", 
      subtitle = "Overhang seats are hollow.") +
 theme_void() +
 theme(
   plot.title = element_text(hjust = 0.5),
   plot.subtitle = element_text(hjust = 0.5),
   legend.position = "none") +
 draw_partylabels(type = "semicircle") +
 scale_colour_manual(
   values = german_data$colour,
   limits = german_data$party_short)
}
\author{
Zoe Meers
}
