% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{geom_parliament_seats}
\alias{geom_parliament_seats}
\title{Parliament seats
The parliament seats geom is used for plotting data from parliament_data().}
\usage{
geom_parliament_seats(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{colour}{ggplot(us_senate, aes(x,y,colour=colour))}
}
\description{
Parliament seats
The parliament seats geom is used for plotting data from parliament_data().
}
\author{
Zoe Meers
}
