% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliament_data.R
\name{parliament_data}
\alias{parliament_data}
\title{A function that prepares data for parliamentary plots}
\usage{
parliament_data(election_data = NULL, parl_rows = NULL,
  party_seats = election_data$seats, total_seats = sum(party_seats),
  party_names = election_data$party, type = c("horseshoe", "semicircle",
  "circle", "classroom"))
}
\arguments{
\item{election_data}{aggregate election results}

\item{parl_rows}{number of rows in parliament}

\item{party_seats}{seats per party}

\item{total_seats}{the total number of seats in parliament}

\item{party_names}{the names of parties elected to parliament}

\item{type}{type of parliament (horseshow, semicircle, circle, classroom, opposing benches)}
}
\description{
A function that prepares data for parliamentary plots
}
\examples{
data <- election_data[which(election_data$year == 2017 & election_data$country == "UK"),]
parl_data <- parliament_data(election_data = data, type = "semicircle", data[[party_seats]] = "seats", data[[party_names]] = "party_short", parl_rows = 6)
ggplot(df, aes(x, y, color=party_long)) + geom_point()

}
\author{
Zoe Meers
}
